fixed $ref issue in NewtonSoft converter
add $ref property when that json object is already seen.
