+++ lib/src/converters/json_converter.dart
@@ -124,50 +124,3 @@ class JsonConverter implements Converter
 	}
 }
 
-class NewtonSoftJsonConverter extends JsonConverter
-{
-	Map<String,String> hashCodesAndTypes = new Map<String,String>();
-
-	@override
-	void setMetaData(Map result, String hashcode, ClassObjectData classObjectData){
-		if(classObjectData.properties.length > 0 )
-		{
-			result["\$id"] = hashcode;
-			setTypeFromObjectType(result, classObjectData);
-		}
-
-		else
-			result["\$ref"] = hashcode;
-	}
-
-	@override
-	String getPreviousHashcode(Map json)
-	{
-		if (json.containsKey("\$ref"))
-			return json["\$ref"];
-
-		return json["\$id"];
-	}
-
-	@override
-	Type findObjectType(dynamic json)
-	{
-		Type objectType = null;
-
-		if (json.containsKey("\$type"))
-			objectType = _getClassMirrorByName( json["\$type"] ).reflectedType;
-		else if
-		(!json.containsKey("\$type") && hashCodesAndTypes.containsKey(json["\$ref"]) )
-			objectType = _getClassMirrorByName( hashCodesAndTypes[json["\$ref"]] ).reflectedType;
-
-		return objectType;
-	}
-
-	@override
-	void afterCreatingClassObjectData(ClassObjectData classObjectData)
-	{
-		if (!hashCodesAndTypes.containsKey(classObjectData.previousHashCode) && classObjectData.properties.containsKey("\$type"))
-			hashCodesAndTypes[classObjectData.previousHashCode] = classObjectData.properties["\$type"].value;
-	}
-
-}
